Execution should not require `Default` implementation for state
Implementing `Default` for state implementations backed by an actual
database is not really straightforward, like: what's a reasonable default
for a database connection pool, or the storage commitment used by a
particular state reader?

It turns out the only place this is used is in `ExecutionEntryPoint::run`,
where a temporary state object is constructed to fetch the initial
OS context and syscall ptr. It turns out that state object is not in
fact required to get these values: the hint processor in the StarknetRunner
was unused for that purpose.

This change moves the `prepare_os_context()` function out to a standalone
function so that we don't have to construct a hint processor (which
requires the initial syscall_ptr) just to get the initial syscall_ptr...