Fix Makefile and README (#322)
* Export CFLAGS and LDFLAGS so they are available to pip install.

Trying to run `make deps` failed on my machine because the environment
variables `CFLAGS` and `LDFLAGS` where not being exported to the
`pip install fastecdsa` command.

In order for _make_ variables to be available to the sub-processes
environment, we need to explicitly use the `export` directive.

* Remove unnecessary venv (de)activation from README

The README instructed the user to activate the python virtualenv before
running `make compile-cairo`, then to deactivate it. This step is
unnecessary since Cairo targets activate the virtual environment inside
the makefile.

* Remove redundant step from README

The step `make build` depends on `make compile-cairo`, so it's not necessary to run the later by hand.