Add missing bits of transaction fee calculation (#305)
* fix(fee): fix cairo_resource_usage validation

There were two flaws in the implementation.

First, the keys in `cairo_resource_usage` has to be a subset of the keys in
`general_config.cairo_resource_fee_weights` and not vice versa.

Second, the `l1_gas_usage` key is a special case and should be ignored.

* fix(constants): add missing resource fee weights

The Python implementation has a default `n_steps` of 0.01 and all other
weights are computed relative to that value.

* fix(fee): use consistent names for cairo builtins

Multiple variants of builtin names were used in
`ExecutionResources.builtin_instance_counter`. `get_execution_resources`
of the Cairo VM returns "bare" names in the hash (like "pedersen"), while
starknet-in-rust code used the variant used by the `cairo-lang` Python
implementation where the builtin name is suffixed by "_builtin".

This change suffixes builtin usage data from the Cairo VM with "_builtin"
so that we use consistent names for fee calculation.

* fixup! fix(fee): fix cairo_resource_usage validation

* fixup! fix(fee): fix cairo_resource_usage validation