DeclareV2 class hash fix (#788)
* Store Sierra class hash in Declare v2 transaction

We'll need this value when we're adding the declared class to storage
so this change moves the computation from the transaction hash calculation
function into DeclareV2::new().

* Fix Declare V2 execution

The code used the _transaction hash_ instead of the Sierra class hash
to add the declared class to storage. This change fixes that to
use the class hash instead.